requirements.txt
pydantic==1.10.4
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.46
